pyproject.toml
[project]
name = "research-assistant"
version = "0.1.0"
description = "AI Research Assistant Agent - 科研助手"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "httpx>=0.27.0",
    "gradio>=4.0.0",
    "python-dotenv>=1.0.0",
    "arxiv>=2.1.0",
]

# Phase 2 依赖（暂不安装）
# langchain>=0.3.0
# langchain-anthropic>=0.3.0
# langgraph>=0.2.0
# chromadb>=0.5.0

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
